requirements.txt
psutil>=5.9.0
matplotlib>=3.6.0
pandas>=1.5.0
numpy>=1.24.0
tqdm>=4.64.0
colorama>=0.4.6
pyyaml>=6.0
flask>=2.3.0
plotly>=5.15.0
cryptography>=41.0.0
rich>=13.0.0
click>=8.1.0
